+++ src/epics.js
@@ -1,0 +1,10 @@
+import { configureEpics } from '@folio/stripes-redux';
+import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/debounceTime';
+import 'rxjs/add/operator/throttleTime';
+import 'rxjs/add/operator/filter';
+import 'rxjs/add/operator/take';
+
+import connectErrorEpic from './connectErrorEpic';
+
+export default configureEpics(connectErrorEpic);