+++ src/index.js
@@ -2,6 +2,8 @@ import 'babel-polyfill';
 import React from 'react';
 import { render } from 'react-dom';
 import { okapi, config } from 'stripes-loader'; // eslint-disable-line
+
+import epics from './epics';
 import configureLogger from './configureLogger';
 import configureStore from './configureStore';
 import { discoverServices } from './discoverServices';
@@ -17,6 +19,6 @@ discoverServices(okapi.url, store);
 const actionNames = gatherActions();
 
 render(
-  <Root store={store} logger={logger} config={config} okapi={okapi} actionNames={actionNames} disableAuth={(config && config.disableAuth) || false} />,
+  <Root store={store} epics={epics} logger={logger} config={config} okapi={okapi} actionNames={actionNames} disableAuth={(config && config.disableAuth) || false} />,
   document.getElementById('root'),
 );